pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openclaw-pg-rag"
version = "1.0.0"
description = "PostgreSQL RAG system for OpenClaw"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "psycopg2-binary>=2.9.0",
    "docling>=2.79.0",
    "xid>=1.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
pg-rag-ingest = "scripts.ingest:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["scripts*", "sql*"]
